Support abstract event classes and common/lib as links so we don't need to reload zone
Currently there is no file watcher for common/lib so you will need to edit a *.cs file in zone folder to get the common/lib changes in to compile but no more reloading of zones! The only time you need to do that is for common/deps which are implementation classes of external 3rd party packages. See AI_NPC and OpenAPI as an example.

Note, common/lib does not compile into common.dll either.